requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jws==0.1.3
MarkupSafe==1.1.1
nltk==3.6.2
numpy==1.20.2
oauth2client==4.1.3
openpyxl==3.0.7
pandas==1.2.4
protobuf==3.16.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase-t==1.0.28
Pyrebase4==4.4.3
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.60.0
urllib3==1.26.4
Werkzeug==1.0.1